[MODIFY] minor code fixes

+++ 2 sem/cpp_basics/lab1/header.h
@@ -5,8 +5,8 @@
 void roundFirst(float&);
 void roundSecond(float*);
 
-void negateFirst(int&);
-void negateSecond(int*);
+void negateFirst(float&);
+void negateSecond(float*);
 
 
 struct square
